Move JsonResponse to Helper.
