+++ src/Helper/JsonResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Lib;
+namespace App\Helper;
 
 use Psr\Http\Message\ResponseInterface as Response;
